requirements.txt
antiorm==1.2.1
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
cfe==0.0.15
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
db==0.1.1
db-sqlite3==0.0.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-countries==7.0
django-crispy-forms==1.6.1
django-environ==0.4.5
django-heroku==0.3.1
django-rest-swagger==2.1.1
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.2
numpy==1.19.5
oauthlib==3.1.0
olefile==0.44
openapi-codec==1.3.2
packaging==20.8
Pillow==8.0.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
SQLAlchemy==1.3.22
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
whitenoise==5.2.0
zipp==3.4.0
